Window closing and Local/Server preference
Take the latest rev of the branch by default (local or server). If the
server tries to load a revision instead of a head (_time = false), then
the comparison will be done anyway with the _saved timestamp of the
local head.

Ask the user if he's sure to quit the window upon closing, only if data
or view file has changed